requirements.txt
chainlit==1.1.404
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-experimental==0.0.64
langchain-google-genai==1.0.10
langchain-ollama==0.1.1
python-dotenv==1.0.1
chromadb==0.5.5
duckduckgo_search==6.2.12
